Clean up and consolidate workflow files
Major cleanup:
- Archived 30+ obsolete files to archive/pre-cleanup-v3/
- Renamed main files to remove version numbers:
  - monitor-enhanced-v3-integrated.js → monitor-enhanced.js
  - automation-enhanced-v3.js → automation-enhanced.js
  - start-enhanced-v3-integrated.sh → start-enhanced-monitor.sh
- Integrated AI attribution system from issue-10
- Updated all references to use new filenames
- Created clean documentation in README-ENHANCED.md

The system is now cleaner and easier to understand:
- One monitor, one automation, one start script
- Clear file naming without version confusion
- All features integrated (AI attribution, MCP monitoring, file org)
