+++ .devcontainer/github-workflow/README-ENHANCED.md
@@ -1,224 +1,3 @@
-# Enhanced GitHub Workflow with ruv-swarm Parallel Processing
+# Enhanced GitHub Workflow Automation
 
-## 🚀 Key Enhancements
-
-### 1. **Dedicated Swarm Processing**
-- Each issue gets its own dedicated swarm for focused analysis
-- Sequential processing ensures each issue receives full attention
-- Claude Code executes the actual work through MCP integration
-
-### 2. **Issue Tracking & Deduplication**
-- Tracks processed tasks in `.processed-tasks.json`
-- Prevents reprocessing of already-handled issues
-- Unique task IDs based on issue number, timestamp, and phase
-
-### 3. **GitHub Projects Support**
-- Supports draft issues (automatically skipped)
-- Label-based filtering (`automation:ignore`, `no-automation`)
-- Project column detection for phase determination
-- Configurable project integration
-
-### 4. **Interactive Flow**
-
-```
-GitHub Issue → Monitor → Interactive Automation → Initialize ruv-swarm
-                ↓                                         ↓
-            Check Labels                    Spawn Coordinator with GitHub Instructions
-                ↓                                         ↓
-            Track Processed                 Spawn Specialized Agents
-                                                         ↓
-                                          Orchestrate with Step-by-Step Updates
-                                                         ↓
-                                          Agents Post Progress to GitHub
-                                                         ↓
-                                          Users Can Guide Process
-```
-
-## 🎯 How It Works
-
-### The Problem We Solved
-The original automation would:
-1. Trigger ruv-swarm but not engage Claude Code
-2. Reprocess the same issues repeatedly
-3. Not support GitHub Projects or filtering
-
-### The Solution
-1. **Interactive Swarm**: Agents post updates to GitHub as they work
-2. **User Guidance**: Users can comment to guide the analysis
-3. **Step-by-Step Updates**: Each agent shares findings via GitHub comments
-4. **Smart Tracking**: Maintains processed task history to prevent reprocessing
-5. **Flexible Filtering**: Skip drafts, respect ignore labels
-6. **Sequential Processing**: Each issue gets dedicated swarm resources
-
-## 📋 Configuration
-
-### Enhanced Config Structure
-```json
-{
-  "filtering": {
-    "ignoreDrafts": true,
-    "ignoreLabels": ["automation:ignore", "no-automation"],
-    "requireLabels": [],
-    "processOnlyPhases": []
-  },
-  "automation": {
-    "parallelBatchSize": 3,
-    "maxConcurrentSwarms": 5
-  }
-}
-```
-
-### Phase-Specific Topologies
-- **Idea**: Mesh topology for exploration
-- **Research**: Hierarchical for structured analysis
-- **Planning**: Hierarchical for organized design
-- **Implementation**: Star for centralized coordination
-
-## 🚀 Quick Start
-
-```bash
-# Set your agent token (for automation use)
-export AGENT_TOKEN="your-github-token-here"
-
-# Start the enhanced monitor
-./start-enhanced-monitor.sh
-
-# Watch the logs
-tail -f monitor.log
-```
-
-## 🐝 Dedicated Swarm Benefits
-
-1. **Focused Analysis**: Each issue gets undivided swarm attention
-2. **Isolated Context**: No cross-contamination between issues
-3. **Resource Optimization**: Each swarm is tailored to the specific issue
-4. **Error Isolation**: Failures don't affect other issues
-
-## 📊 Monitoring & Tracking
-
-### Files Created
-- `.monitor-state.json` - Monitor state and ETags
-- `.processed-tasks.json` - History of processed tasks
-- `.swarm-context-{issue}.json` - Temporary context for Claude
-
-### Status Indicators
-- ✅ Processed successfully
-- ⏭️ Skipped (filtered)
-- ✓ Already processed (cached)
-- ❌ Processing error
-
-## 🏷️ Label-Based Control
-
-### Skip Processing
-Add any of these labels to skip automation:
-- `automation:ignore`
-- `no-automation`
-- `wip`
-
-### Phase Labels
-- `phase:idea`
-- `phase:research`
-- `phase:planning`
-- `phase:implementation`
-
-### Automation Control Labels
-- `auto-close-on-complete` - Auto-close issue when implementation phase completes
-- `keep-open` - Never auto-close this issue (overrides auto-close)
-
-### Status Labels (Added Automatically)
-- `swarm-in-progress` - Added when swarm starts, prevents duplicate processing
-- `swarm-processed` - Replaces in-progress label when analysis completes
-
-## 🔧 Troubleshooting
-
-### Monitor Won't Start
-1. Check GitHub token is set
-2. Verify config-enhanced.json exists
-3. Look for port conflicts
-
-### Issues Not Processing
-1. Check if issue has ignore labels
-2. Verify issue isn't a draft
-3. Check `.processed-tasks.json` for history
-
-### No Results Posted
-1. Ensure Claude Code has both ruv-swarm and GitHub MCP tools configured
-2. Check that AGENT_TOKEN is set in environment
-3. Verify token has write permissions to the repo
-4. Monitor console output for Claude instructions
-5. Claude should use:
-   - `mcp__ruv-swarm__*` tools for coordination
-   - `mcp__github__*` tools for posting results
-
-## 🎯 Best Practices
-
-1. **Use Phase Labels**: Explicitly label issues for better routing
-2. **Batch Similar Issues**: Group related issues for processing
-3. **Monitor Resources**: Watch for rate limits
-4. **Clean History**: Periodically clean `.processed-tasks.json`
-
-## 🚀 Advanced Features
-
-### Custom Filtering
-```javascript
-// In config-enhanced.json
-"filtering": {
-  "requireLabels": ["ready-for-automation"],
-  "processOnlyPhases": ["research", "planning"]
-}
-```
-
-### Project Integration
-```javascript
-// In config-enhanced.json
-"github": {
-  "projectNumber": 1,
-  "projectColumnId": "column_123"
-}
-```
-
-### Processing Configuration
-```javascript
-// In config-enhanced.json
-"automation": {
-  "processingDelay": 2000  // 2 second delay between issues
-}
-```
-
-## 🤖 How the Interactive Swarm Works
-
-The automation creates an interactive analysis process:
-
-1. **Initialize Swarm**: Sets up hierarchical topology for coordination
-2. **Spawn Coordinator**: With explicit instructions to post GitHub updates
-3. **Spawn Specialists**: Each agent knows to coordinate through GitHub
-4. **Orchestrate Task**: Includes instructions for step-by-step updates
-5. **Interactive Updates**: Agents use Claude to post progress via GitHub MCP
-6. **User Participation**: Users can comment to guide the analysis
-7. **Final Synthesis**: Coordinator posts comprehensive findings
-
-Key Instructions Given to Agents:
-- "Use Claude to post updates to GitHub issue #X"
-- "Tell Claude to use mcp__github__add_issue_comment"
-- "Post updates for each major discovery"
-- "Check for user feedback and incorporate it"
-
-## 💬 Interactive Analysis Benefits
-
-The interactive approach provides:
-- **Transparency**: See what agents are thinking in real-time
-- **User Control**: Guide the analysis with your comments
-- **Progressive Updates**: No waiting for final results
-- **Collaborative**: Agents and users work together
-- **Context-Aware**: Agents can adjust based on feedback
-
-## 🐝 The Power of ruv-swarm
-
-This enhanced automation leverages:
-- **Interactive Agents**: Post updates as they work
-- **GitHub Integration**: Direct posting via Claude's MCP tools
-- **Sequential Processing**: Each issue gets its own dedicated swarm
-- **Smart Coordination**: Agents collaborate through GitHub
-- **User Engagement**: Analysis becomes a conversation
-
-Together, they create a powerful GitHub automation system that processes issues interactively and transparently!
+A comprehensive GitHub automation system with AI attribution, MCP monitoring, and intelligent file organization.
