+++ .devcontainer/github-workflow/archive/pre-cleanup-v3/automation-swarm-enhanced.js
