+++ .devcontainer/github-workflow/archive/pre-cleanup-v3/automation.js
