+++ .devcontainer/github-workflow/archive/pre-cleanup-v3/mcp-server-monitor.js
