+++ .devcontainer/github-workflow/archive/pre-cleanup-v3/start-coordinator-monitor.sh
