+++ .devcontainer/github-workflow/archive/pre-cleanup-v3/test-claude-github-simple.sh
