+++ .devcontainer/github-workflow/archive/pre-cleanup-v3/test-claude-integrated.js
